% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermitage.R
\docType{data}
\name{edu}
\alias{edu}
\title{example data: Access to basic education dataset from https://ourworldindata.org/children-not-in-school}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4329 rows and 4 columns.
}
\usage{
edu
}
\description{
example data: Access to basic education dataset from https://ourworldindata.org/children-not-in-school
}
\keyword{datasets}
