% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermitage.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{example data: GDP dataset from https://ourworldindata.org/global-economic-inequality-introduction}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19878 rows and 4 columns.
}
\usage{
gdp
}
\description{
example data: GDP dataset from https://ourworldindata.org/global-economic-inequality-introduction
}
\keyword{datasets}
