% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermitage.R
\name{hermitage_palette}
\alias{hermitage_palette}
\title{Discover palette}
\format{

}
\usage{
hermitage_palette("parsons_2")
hermitage_palette("hermitage_1")
hermitage_palette("harmony", 5)
hermitage_palette("imari", 5)
hermitage_palette("du_barry", 3)
For more colours:
more_colours <- hermitage_palette(25, name = "imari", type = "continuous")
}
\arguments{
\item{name}{Name of the palette. Options:
 \code{parsons_1}, \code{parsons_2},  \code{du_barry},
 \code{delft_ware}, \code{imari},  \code{collioure_matisse}, \code{peacock_clock},
 \code{prodigal_son},  \code{harmony} , \code{dance_matisse} ,
 \code{judith}, \code{magdalene_titian}, \code{bagration}, \code{flight_gauguin},
 \code{hermitage_1}, \code{hermitage_2}, \code{madonna_litta}, \code{cottages_vincent},
\code{danae}, \code{battista_cima}, \code{weyden}, \code{faberge},
\code{duresco}, \code{water_paint}, \code{kunstnefarver}, \code{klee}.}

\item{n}{Number of colours needed. The largest palettes are parsons_1 and parsons_2 inspired by Thomas Parsons's colour samples from the book "A Tint Of Historical colours" (1934) and include 87 colours and 120 colours, respectively.
du_barry (n=4), delft_ware (n=4), imari (n=7) palettes and several others are inspired by the book "The anatomy of colour" by Patrick Baty. Other palettes are inspired by art collection of Hermitage museum.}

\item{type}{Either "continuous" or "discrete".}
}
\value{
A vector of colours
}
\description{
Discover palette
}
\examples{
hermitage_palette("parsons_1")
}
\keyword{colours}
