% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermitage.R
\docType{data}
\name{hermitage_palettes}
\alias{hermitage_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 27.
}
\usage{
hermitage_palettes
}
\description{
Use \code{\link{hermitage_palette}} to define palettes of needed size/length
}
\keyword{datasets}
